FE: Move login state route guard into `beforeCreate` hook of `App.vue`
